% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bivariate_pdf.R
\name{make_bivariate_pdf}
\alias{make_bivariate_pdf}
\title{Estimates a bivariate empirical density function}
\usage{
make_bivariate_pdf(x, y, w = NULL)
}
\arguments{
\item{x}{Numeric. A continuous variable.}

\item{y}{Numeric. A continuous variable.}

\item{w}{Numeric. Frequency Weights.}
}
\value{
Returns a bivariate function which evaluates to the empirical joint distribution
}
\description{
Given two continuous variables, this function estimates their empirical joint
distribution via kernel density smoothing.
}
