% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_CA.R
\name{make_CA}
\alias{make_CA}
\title{Returns a function which calculates the Comparative Advantage (CA) between 2 income levels}
\usage{
make_CA(reg, income_var, performance_var, good_perfomance = NULL)
}
\arguments{
\item{reg}{a glm binary logistic regression object}

\item{income_var}{A character vector with the name of the income variable,
as it appear in \code{all.vars(formula(reg))}.}

\item{performance_var}{A character vector with the name of the schooling
performance variable, as it appear in \code{all.vars(formula(reg))}.}

\item{good_perfomance}{Numeric.}
}
\value{
Returns a bivariate function which calculates the comparative advantage
  between two points/observations on the income distribution.
}
\description{
Returns a function which calculates the Comparative Advantage (CA) between 2 income levels
}
