% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_CA_meanEffect.R
\name{make_CA_meanEffect}
\alias{make_CA_meanEffect}
\title{Returns a function which calculates Mean Effect of the Comparative Advantage (CA)}
\usage{
make_CA_meanEffect(reg, income_var, performance_var, good_perfomance = NULL)
}
\arguments{
\item{reg}{a glm binary logistic regression object}

\item{income_var}{A character vector with the name of the income variable,
as it appear in \code{all.vars(formula(reg))}.}

\item{performance_var}{A character vector with the name of the schooling
performance variable, as it appear in \code{all.vars(formula(reg))}.}
}
\value{
Returns a bivariate function which calculates the CA Mean effect
  between two points/observations on the income distribution.
}
\description{
Returns a function which calculates Mean Effect of the Comparative Advantage (CA)
}
