% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_conditional_pdf.R
\name{make_conditional_pdf}
\alias{make_conditional_pdf}
\title{Estimates a empirical conditional density function}
\usage{
make_conditional_pdf(x, given_y, w = NULL)
}
\arguments{
\item{x}{Numeric. A continuous variable.}

\item{given_y}{Numeric. A continuous variable on which the distribution of \code{x} will be conditioned on}

\item{w}{Numeric. Frequency Weights.}
}
\value{
Returns a bivariate function which evaluates to the empirical distribution of \code{x} conditional on \code{given_y}
}
\description{
Given two continuous variables, this function estimates their empirical conditional
distribution via kernel density smoothing.
}
