% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_INE.R
\name{make_INE}
\alias{make_INE}
\title{Returns a function which calculates the Educational Inequality (INE) between 2 income levels}
\usage{
make_INE(reg, income_var, performance_var)
}
\arguments{
\item{reg}{a glm binary logistic regression object}

\item{income_var}{A character vector with the name of the income variable,
as it appear in \code{all.vars(formula(reg))}.}

\item{performance_var}{A character vector with the name of the schooling
performance variable, as it appear in \code{all.vars(formula(reg))}.}
}
\value{
Returns a bivariate function which calculates the difference in
  probability of a successful educational transition (as modelled by \code{reg})
  between two points/observations on the income distribution. Schooling perfomance
  is already integrated out and all covariated set to their respective mean.
}
\description{
Returns a function which calculates the Educational Inequality (INE) between 2 income levels
}
