% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{Mode (most frequent value) of a distribution}
\usage{
mode(x, w = NULL)
}
\arguments{
\item{x}{A numeric, character, or factor vector.}

\item{w}{Numeric. Frequency weights}
}
\value{
Returns a single-valued vector of the same class as x.
}
\description{
Identifies and returns the most frequent value of a variable.
}
